mcp/requirements.txt
google-adk==0.0.18
mcp==0.1.6
fastapi==0.110.0
uvicorn==0.29.0
